+++ changelog.md
@@ -8,7 +8,7 @@
 - Updated dev dependencies, some of which require newer Node.js versions than were previously supported.
 - The API is now ESM in `.mjs` files instead of CJS in `.js` files, [accessible via `import` but not `require`](https://nodejs.org/dist/latest/docs/api/esm.html#esm_require).
 - The function `installFrom` is now a default export.
-- The tests are now ESM in a `.mjs` file instead of CJS in a `.js` file.
+- The tests are now ESM in `.mjs` files instead of CJS in a `.js` file.
 
 ### Patch